+++ R/zzz_data.R
@@ -98,6 +98,10 @@ pride_data <- list(
     colors = c("#FFB187", "#FF7D51", "#FFF7B8", "#000000", "#B5B5B5"),
     theme = theme_ggplot2
   ),
+  stellarian = list(
+    colors = c("#B6D6C1", "#79C49D", "#FFF7B8", "#000000", "#B5B5B5"),
+    theme = theme_ggplot2
+  ),
   # Other genders
   genderfluid = list(
     colors = c("#FF76A4", "#FFFFFF", "#C011D7", "#000000", "#2F3CBE")