+++ NEWS.md
@@ -1,5 +1,6 @@
 # gglgbtq 0.1.1
 
+* Fixed potential error due to directly accessing `ggplot2::theme()`.
 * Implemented the following palettes: `agender`, `alloace`, `aroallo`, `bigender`, `demiboy`, `demigirl`.
 * Changed `gay_man` to `gay` (and, analogously, `gay_7`).
 * Moved palettes to the "Gallery" vignette.