+++ NEWS.md
@@ -1,4 +1,4 @@
-# gglgbtq (development version)
+# gglgbtq 0.2.0
 
 * Added `scale_color_lgbtq()` and `scale_fill_lgbtq()` for convenience.
 * Implemented the following palettes: `acesthetic`, `analterous`, `asensual`, `ace_spectrum`, `aro_spectrum`, `aroace`, `androsexual`, `gynesexual`, `abrosexual`, `helian`, `lunarian`, `solarian`, `stellarian`, `amatopunk`, `bear`, `butch`, `femme`, `otter`, `queerhet`.